pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3", "wheel"]
build-backend = "setuptools.build_meta"
[project]
name = "hypixelez"
description = "Python library for easy interaction with Hypixel API, specifically designed for SkyBlock profile data"
keywords = ["hypixel", "minecraft", "skyblock", "api", "gaming"]
readme = {file = "README.md", content-type = "text/markdown"}
license = {text = "MIT"}
authors = [
  {name = "Onii-Chan", email = "skobochki.ad@mail.ru"}
]
version = "1.0.0"

requires-python = ">=3.9"
classifiers = [
  "Development Status :: 5 - Production/Stable",

  "Intended Audience :: Developers",
  "Intended Audience :: Education",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Games/Entertainment",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
]

dependencies = [
    "requests>=2.32.5"
]

[project.urls]
Homepage = "https://github.com/SerJo2/hypixelez"
Repository = "https://github.com/SerJo2/hypixelez"
Issues = "https://github.com/SerJo2/hypixelez/issues"
Changelog = "https://github.com/SerJo2/hypixelez/blob/main/CHANGELOG.md"

[tool.setuptools]
packages = {find = {where = ["src"], exclude = ["tests*", "tests.*"]}}
package-dir = {"" = "src"}

[project.optional-dependencies]
test = [
    "pytest>=6.0",
    "pytest-cov",
    "python-dotenv"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]